pyproject.toml
[project]
name = "atumm-service-user"
version = "0.1.0"
description = "User Webservice"
authors = [
    {name = "Omars", email = "omar@atumm.org"},
]
dependencies = [
    "pydantic>=2.3.0",
    "py-bcrypt>=0.4",
    "injector>=0.21.0",
    "beanie>=1.21.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.1",
    "pytest-cov>=4.1.0",
    "faker>=19.3.1",
    "mock>=5.1.0",
    "ssort>=0.11.6",
    "isort>=5.12.0",
    "black>=23.7.0",
    "anyio[asyncio]>=4.0.0",
    "bump-pydantic>=0.7.0",
]
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
